requirements.txt
openai
google-generativeai
langchain
langchain-google-genai
aiohttp==3.10.11
numpy>=1.26.0
python-binance==1.0.17
python-dotenv==1.0.0
requests==2.32.0
ta==0.10.2
websockets==11.0.3
fastapi
uvicorn
apscheduler
discord
nest_asyncio
pandas>=2.1.0
schedule
colorlog